-requirements.txt
accelerate==1.8.1
certifi==2025.6.15
charset-normalizer==3.4.2
diffusers==0.34.0
filelock==3.13.1
fsspec==2024.6.1
hf-xet==1.1.5
huggingface-hub==0.33.1
idna==3.10
importlib_metadata==8.7.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==2.1.2
packaging==25.0
peft==0.15.2
pillow==11.0.0
psutil==7.0.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
setuptools==70.2.0
sympy==1.13.3
tokenizers==0.21.2
torch==2.7.1
torchaudio==2.7.1
torchvision==0.22.1
tqdm==4.67.1
transformers==4.53.0
typing_extensions==4.12.2
urllib3==2.5.0
zipp==3.23.0
